requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
drf-yasg==1.21.7
pandas==2.0.3
yfinance==0.2.28
numpy==1.24.3
scipy==1.11.1
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.3.0
matplotlib==3.7.2
torch==2.3.1
